requirements.txt
# Backend
fastapi
uvicorn[standard]

# ML & AI
transformers
sentence-transformers
torch
chromadb

# Data Processing
pymupdf
beautifulsoup4
langchain
langchain-text-splitters
python-dotenv

# Testing
pytest
anyio # For async testing with pytest
